Use generators Degree for better unimod detection
1) There is a bool potentially_unimodular in simplex.cpp. It is true if and only if height <=1 and the gcd of the degrees of the generators is 1.
I hope this helps to recognize nonunimodular simplices of height 1.

2) key_t is now unsigned long. We know that we are on the safe side with it.

3) recursion_allowed=frue inserted for -s.

+++ libnormaliz/libnormaliz.h
@@ -61,7 +61,7 @@ using Mode::ComputationMode;
 
 /* this type is used in the entries of keys
  * it has to be able to hold number of generators*/
-typedef unsigned short key_t;
+typedef unsigned long key_t;
 
 extern bool verbose;
